% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{bvq_connect}
\alias{bvq_connect}
\title{Authenticate in Google and formr}
\usage{
bvq_connect(google_email = NULL, verbose = TRUE)
}
\arguments{
\item{google_email}{E-mail used in Google Drive account. If NULL (default), it is assumed to be the same as formr_email.}

\item{verbose}{Should progress messages and warnings be printed in the console}
}
\value{
Logical. TRUE if Google and formr authentication was successful, FALSE if authentication of any of the two failed.
}
\description{
Authenticate in Google and formr
}
\details{
This function tries to log in to the formr API by trying to retrieve its corresponding key via the keyring package. If no key exists under the name "multilex", the user is prompted to create it first.
}
\examples{
my_email <- "gonzalo.garciadecastro@upf.edu"
bvq_connect(google_email = my_email)
}
