% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/items.R
\name{bvq_items}
\alias{bvq_items}
\title{Get items included in a questionnaire}
\usage{
bvq_items(section, version = "bvq-1.0.0")
}
\arguments{
\item{section}{Name of the version of the questionnaire for which the items of a section will be retrieved. Check the output of \code{names(get_bvq_runs())} to see the available versions of the questionnaire.}
}
\value{
A list of length 3, which includes:
\itemize{
\item survey: A \link[tibble:tibble]{tibble::tibble} containing the items included in the questionnaire and several properties. Each row corresponds to a single item, and each column corresponds to a particular property:
\itemize{
\item type: a character string indicating the type of the item (see \href{https://formr.org/documentation#available_items}{formr documentation}).
\item name: a character string indicating the name of the item, as it appears in the output of \code{\link[=bvq_responses_vocab]{bvq_responses_vocab()}}.
\item label: a character string indicating the text shown to participants when filling out the questionnaire.
\item optional: a logical value indicating whether providing an answer to the item is mandatory for participants.
\item class: a character string indicating the CSS class with of the item.
\item showif: a character string indicating R code that determines under what conditions the item is shown to participants.
\item value: default value of the item.
\item block_order: character string (a latter) indicating the order in which the block that the item belongs to appears in the survey.
\item item_order: integer indicating the order in which the item appears within the block it belongs to belongs.
}
\item choices: A \link[tibble:tibble]{tibble::tibble} containing the choices given to participants for some items. Each row corresponds to a choice, and each column corresponds to a particular choice property:
\itemize{
\item list_name`: character string indicating the name of the name of the choice list (which may repeat across different items).
\item name: character string indicating the name that a particular choice will be assigned in the code.
\item label: character string indicating the text that will be show to participants for a particular choice.
}
\item settings: A \link[tibble:tibble]{tibble::tibble} containing the settings for the survey. Each row corresponds to one setting, and each column indicates the setting names and values:
\itemize{
\item item: name of the setting.
\item value: value of the setting.
}
}
}
\description{
This function retrieves information about the items in a particular section in the BVQ questionnaire. This includes item names, item types, text, choices, settings, and other metadata.
}
\examples{
\dontrun{
bvq_items("bvq_06_words_catalan", version = "bvq-1.0.0")
}

}
\author{
Gonzalo Garcia-Castro
}
