% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{bvq_logs}
\alias{bvq_logs}
\title{Generate participant information and progress for each response}
\usage{
bvq_logs(
  participants = NULL,
  responses = NULL,
  bilingual_threshold = 0.8,
  other_threshold = 0.1,
  verbose = TRUE
)
}
\arguments{
\item{participants}{Participants data frame, as generated by
\code{\link{bvq_participants}}. If NULL (default),
\code{\link{bvq_participants}} is run.}

\item{responses}{Responses data frame, as generated by
\code{\link{bvq_responses}}. If NULL (default), \code{\link{bvq_responses}}
is run.}

\item{bilingual_threshold}{Numeric scalar ranging from 0 to 1 indicating the
minimum degree of exposure to Catalan or Spanish to consider a participant
as
\emph{Monolingual}.}

\item{other_threshold}{Numeric scalar ranging from 0 to 1 indicating the
minimum degree of exposure to languages other than Catalan and Spanish to
consider a participant as \emph{Other}.}

\item{verbose}{Should progress messages and warnings be printed in the
console}
}
\value{
A data frame (actually, a \code{\link[tibble]{tibble}}) with
participant-level information. Each row corresponds to a questionnaire
response and each column represents a variable. The output includes the
following variables:
\describe{
\item{id}{a character string indicating a participant's identifier. This value is always the same for each participant, so that different responses from the same participant share the same \code{id}.}
\item{id_exp}{a character string indicating a participant's identifier in the context of the particular study in which the participant was tested and invited to fill in the questionnaire. This value is always the same for each participant within the same study, so that different responses from the same participant in the same study share \code{id_exp} The same participant may have different \code{id_exp} across different studies.}
\item{id_db}{a character string with five digits indicating a participant's identifier in the database from the \href{https://www.upf.edu/web/cbclab}{Laboratori de Recerca en Infància} at Universitat Pompeu Fabra. This value is always the same for each participant, so that different responses from the same participant share the same \code{id_db}.}
\item{code}{a character string identifying a single response to the questionnaire. This value is always unique for each response to the questionnaire, even for responses from the same participant.}
\item{time}{a numeric value indicating how many times a given participant has been sent the questionnaire, regardless of whether they completed it or not.}
\item{study}{a character string indicating the study in which the participant was invited to fill in the questionnaire. Frequently, participants that filled in the questionnaire came to the lab to participant in a study, and were then invited to fill in the questionnaire later. This value indicates what study each participant was tested in before being sent the questionnaire.}
\item{version}{a character string indicating what version of the questionnaire a given participant filled in. Different versions may contain a different subset of items, and the administration instructions might vary slightly (see formr questionnaire templates in the \href{https://github.com/gongcastro/multilex}{GitHub repository}). Also, different versions were designed, implemented, and administrated at different time points (e.g., before/during/after the COVID-related lockdown).}
\item{date_sent}{a date value (see lubridate package) in \code{yyyy/mm/dd} format indicating the date in which the questionnaire was sent to participants.}
\item{days_from_sent}{a numeric value indicating the number of days elapsed since participants were sent the questionnaire (as indicated by \code{date_sent})  and completed the questionnaire.}
\item{date_birth}{a date value (see lubridate package) in \code{yyyy/mm/dd} format indicating participants birth date.}
\item{age}{a numeric value indicating the number of months elapsed since participants' birth date until they filled in the last item of their questionnaire response.}
\item{age_today}{a numeric value indicating the number of months elapsed since participants' birth date until the present day, as indicated by \code{\link[lubridate]{now}.}}
\item{months_from_last_response}{a numeric value indicating the number of months elapsed since participants' last questionnaire response (as indicated  by \code{time_stamp}) until the present day, as indicated by \code{\link[lubridate]{now}.}}
\item{sex}{a character string indicating participants' biological sex, as reported by the parents.}
\item{postcode}{a character string indicating participants' household postcode.}
\item{edu_parent1}{a character string indicating the educational attainment of one of the parents/caretakers.}
\item{edu_parent2}{a character string indicating the educational attainment of the other parent/caretaker, if any.}
\item{dominance}{a character string indicating the language of highest exposure ("Catalan" or "Spanish"), as reported by parents. If exposure is identical for both language, "Catalan" is assigned.}
\item{lp}{a character string indicating participants' language profile, classified using parental reports of language exposure (see \code{doe_spanish}, \code{doe_catalan}, and \code{doe_others}), and the thresholds passed in the \code{bilingual_threshold} and \code{other_threshold}.}
\item{doe_spanish}{a numeric value ranging from 0 to 1 indicating participants' daily exposure to Spanish, as estimated by parents/caretakers. This value aggregates participants' exposure to any variant of Spanish (e.g., European and American Spanish).}
\item{doe_catalan}{a numeric value ranging from 0 to 1 indicating participants' daily exposure to Catalan, as estimated by parents/caretakers. This value aggregates participants' exposure to any variant of Catalan (e.g., Catalan from Majorca or Barcelona).}
\item{doe_others}{a numeric value ranging from 0 to 1 indicating participants' daily exposure to languages other than Spanish or Catalan, as estimated by parents/caretakers, aggregating participants' exposure to all those other languages (e.g., Norwegian, Arab, Swahili).}
\item{progress}{a numeric value ranging from 0 to 1 indicating participants' progress filling the questionnaire. A value of 0 indicates that the participant has not filled in any item yet. A value of 0.5 indicates that the participant is halfway through the questionnaire. A value of 1 indicates that the participant has completed all items.}
\item{completed}{a logical value that returns TRUE if \code{progress} is 1, and FALSE otherwise.}
}
}
\description{
This function generates a data frame that contains
participant-level information. Each row is a given participant's response
and each column is a variable. The same participant will always be
identified with the same \code{id}. The variable \code{time} indexes how
many times a participant has been sent the questionnaire, independently of
whether a response was obtained from them later.
}
\author{
Gonzalo Garcia-Castro
}
