% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{bvq_norms}
\alias{bvq_norms}
\title{Generate item-level norms for age, sex, language profile and item dominance}
\usage{
bvq_norms(
  participants,
  responses,
  item = NULL,
  language = c("Catalan", "Spanish"),
  type = c("understands", "produces"),
  age = NULL,
  lp = c("Bilingual", "Monolingual", "Other"),
  sex = c("Female", "Male"),
  semantic_category = NULL,
  .width = 0.95,
  ...
)
}
\arguments{
\item{participants}{Participants data frame, as generated by
\code{\link[=bvq_participants]{bvq_participants()}}.}

\item{responses}{Responses data frame, as generated by \code{\link[=bvq_responses]{bvq_responses()}}.}

\item{item}{Character string indicating the item to compute norms for. If
left \code{NULL} (default) norms will be computed for all items. You can check
the available items in the \link{pool} data set running \code{data("pool")}.}

\item{language}{Character string indicating the language to compute
vocabulary norms for: \code{"catalan"} and/or \code{"spanish"}.}

\item{type}{Character string indicating the vocabulary type to compute norms
for. Takes \code{"understands"} and/or \code{"produces"} (defaults to both).}

\item{age}{Numeric vector of length two (min-max) indicating the age range of
participants to compute norms for.}

\item{lp}{character string indicating the language profile of participants to
compute norms for: \code{"Bilingual"}, \code{"Monolingual"}, \code{"Other"} (defaults to
all).}

\item{sex}{character string indicating the sex of participants to compute
norms for. Takes \code{"Female"} and/or \code{"Male"} (defaults to both).}

\item{semantic_category}{character string indicating the semantic/functional
category or categories to include items from. See available categories in
the \link{pool} data set by running \code{data("pool")}.}

\item{.width}{Numeric values ranging from 0 to 1 (not included) indicating
the confidence level of confidence intervals (defaults to \code{0.95}).}

\item{...}{Unused}
}
\value{
A data frame (actually, a \link[tibble:tibble]{tibble::tibble} with the proportion of
participants in the sample that understand or produce the items indicated
in \code{item}, along with the standard error and confidence interval of the
estimation. The output contains the following variables:
\itemize{
\item te: an integer identifying the Translation Equivalent (a.k.a., pair of cross-language synonyms, doublets) the item belongs to.
\item item: character string indicating the item identifier (e.g., \code{spa_mesa}). This value is unique for each item. Responses to the same item from different participants are linked by the same \code{item} value.
\item language: a character string indicating the language the item response belongs to: \code{"Catalan"} if item in Catalan), \code{"Spanish"} if item in Spanish.
\item age: an integer indicating the age of participants (in months) for which the estimates should be computed. If a non-integer is provided (e.g., \code{15.36}, it is rounded downwards using \code{\link[=floor]{floor()}}.)
\item type: a character string indicating the vocabulary type computed: \code{"understands"} if option 'Understands' was selected, and \code{"produces"} if option 'Understands & Says' was selected.
\item lp: a character string indicating participants' language profile, classified using parental reports of language exposure (see \code{doe_spanish}, \verb{"doe_catalan}, and \code{doe_others}), and the thresholds passed in the \code{bilingual_threshold} and \code{other_threshold}.
\item semantic_category: a character string indicating the semantic/function category the item belongs to (e.g., \code{"Vehicles"}, \code{"Actions"}).
\item item_dominance: a character string that takes the value \code{"L1"} if the item belongs to participants' language of most exposure, and L2 if the item belongs to participants' language of least exposure.
\item label: a character string indicating the text presented to participants in the questionnaire (replacing the \code{item} identifier).
\item .sum: a positive integer indicating the number of positive responses: \code{responses} is 2 (Understands) or 3 (Understands & Says) for \code{type = 'understands'}, and 3 (Understands & Says) if \code{type = 'produces'}.
\item .n: a positive integer indicating the total number number of responses (useful for computing proportions).
\item .prop: a numeric value ranging from 0 to 1 (both included) indicating the estimated proportion of participants that provided a positive response, adjusted following Gelman et al.'s method to account for zero- and one-inflation (see function \link{prop_adj}).
}
}
\description{
This function generates a data frame with the estimated proportion of
children that understand and/or produce some items for a selected age range
and participant profiles. Estimated proportions and corresponding standard
errors and confidence intervals are computed adjusting for zero- and
one-inflation (see function \code{\link[=prop_adj]{prop_adj()}}).
}
\author{
Gonzalo Garcia-Castro
}
