% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{bvq_vocabulary}
\alias{bvq_vocabulary}
\title{Generate participant information and progress for each response}
\usage{
bvq_vocabulary(
  participants = bvq_participants(),
  responses = bvq_responses(participants),
  ...,
  .scale = "prop"
)
}
\arguments{
\item{participants}{Participants data frame, as generated by
\code{\link[=bvq_participants]{bvq_participants()}}.}

\item{responses}{Responses data frame, as generated by
\code{\link[=bvq_responses]{bvq_responses()}}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Unquoted name of the
variable(s) to group data into. Vocabulary metrics will be calculated by
aggregating responses within the groups that result from the combination of
crossing of the variables provided in \code{...}. These variables can refer to
item properties (see \link{pool}, e.g., \code{semantic_category}) or to
participant properties (see \code{\link[=bvq_logs]{bvq_logs()}}, e.g., \code{lp}).}

\item{.scale}{A character vector that takes the value \code{"count"} and/or
\code{"prop"}. If \code{"prop"} (default), vocabulary metrics are calculated as
proportions. If \code{"count"}, vocabulary metrics are reported as counts
(number of words).}
}
\value{
A dataset (actually, a \link[tibble:tibble]{tibble::tibble} with each participant's
comprehensive and/or vocabulary size in each language. This data frame
contains the following variables:
\itemize{
\item child_id: a character string with five digits indicating a participant's identifier in the database from the \href{https://www.upf.edu/web/cbclab}{Laboratori de Recerca en Infància} at Universitat Pompeu Fabra. This value is always the same for each participant, so that different responses from the same participant share the same \code{child_id}.
\item time: a numeric value indicating how many times a given participant has been sent the questionnaire, regardless of whether they completed it or not.
\item age: a numeric value indicating the number of months elapsed since participants' birth date until they filled in the last item of their questionnaire response.
\item type: a character string indicating the vocabulary type computed: \code{"understands"} if option "Understands" was selected, and \code{"produces"} if option "Understands & Says" was selected.
\item total_count: integer indicating the number of items selected as "Understands" or "Understands and Says" in both languages.
\item l1_count: positive integer indicating the number of items selected as "Understands" or "Understands and Says" in the dominant language (L1).
\item l2_count: positive integer indicating the number of items selected as "Understands" or "Understands and Says" in the non-dominant language (L2).
\item concept_count: positive integer indicating the number of translation equivalents (a.k.a. cross-language synonyms or doublets) in which  "at list one of the items was selected as "Understands" or "Understands and Says". This is a measure of the number of lexicalised concepts.
\item te_count: positive integer indicating the number of translation equivalents (out of the total number of items the participant answered to) in which at both items was selected as "Understands" or "Understands and Says". This is a measure of the number of lexicalised concepts.
\item total_prop: numeric value ranging from 0 to 1 (both included) indicating the proportion of items selected as "Understands" or "Understands and Says" in both languages.
\item l1_prop: numeric value ranging from 0 to 1 (both included) indicating the proportion of of items selected as "Understands" or "Understands and Says" in the dominant language (L1).
\item l2_prop: numeric value ranging from 0 to 1 (both included) indicating the proportion of of items selected as "Understands" or "Understands and Says" in the non-dominant language (L2).
\item concept_prop: numeric value ranging from 0 to 1 (both included) indicating the proportion of of translation equivalents (a.k.a. cross-language synonyms or doublets) in which at least one of the items was selected as "Understands" or "Understands and Says". This is a measure of the number of lexicalised concepts.
\item te_prop: numeric value ranging from 0 to 1 (both included) indicating the proportion of of translation equivalents (aka. cross-language synonyms or doublets) in which both items were selected as "Understands" or "Understands and Says". This is a measure of the number of lexicalised concepts.
The specific subset of columns returned by \code{\link[=bvq_vocabulary]{bvq_vocabulary()}} depends
on the elements of \code{...} and \code{.scale}.
}
}
\description{
This function generates a data frame with the vocabulary of each participant
(keeping longitudinal data from the same participant in different rows).
Comprehensive and productive vocabulary sizes are computed as raw counts
(\verb{*_count}) and as proportions \verb{*_prop}.
}
\author{
Gonzalo Garcia-Castro
}
