% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{check_arg_dots}
\alias{check_arg_dots}
\title{Check argument \code{...} in the \code{\link[=bvq_vocabulary]{bvq_vocabulary()}} function}
\usage{
check_arg_dots(x, .cols)
}
\arguments{
\item{x}{Variable names passed with \code{...}}

\item{.cols}{Valid column names}
}
\description{
Check argument \code{...} in the \code{\link[=bvq_vocabulary]{bvq_vocabulary()}} function
}
