% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{check_arg_te}
\alias{check_arg_te}
\title{Check argument \code{te} in \code{\link[=bvq_norms]{bvq_norms()}} function}
\usage{
check_arg_te(te, item)
}
\arguments{
\item{te}{Translation equivalent for which the norms should be computed.
\itemize{
\item If NULL (default), norms are computed exclusively for the items indicated in \code{item}.
\item If TRUE, norms are computed for both the item indicated in the \code{item} argument, and for its translation.
\item If FALSE, norms are computed exclusively for the items indicated in \code{item} (same as NULL)
\item If numeric vector, norms are computed for all items corresponding to the translation equivalents indicated in this argument.
}}

\item{item}{Character string indicating the item to compute norms for. If
left \code{NULL} (default) norms will be computed for all items. You can check
the available items in the \link{pool} data set running \code{data("pool")}.}
}
\description{
Check argument \code{te} in \code{\link[=bvq_norms]{bvq_norms()}} function
}
