% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{coalesce_by_column}
\alias{coalesce_by_column}
\title{Fill missing with previous row}
\usage{
coalesce_by_column(x)
}
\arguments{
\item{x}{Vector whose missing values will be filled with parallel non-missing
values}
}
\description{
Fill missing with previous row
}
