% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_in_months}
\alias{diff_in_months}
\title{Age difference in months}
\usage{
diff_in_months(x, y)
}
\arguments{
\item{x}{Most recent date}

\item{y}{Least recent date}
}
\value{
Absolute difference in months elapsed between \code{x} and \code{y}
}
\description{
Returns the absolute difference in months elapsed between two dates, as indicated by \code{\link[lubridate:time_length]{lubridate::time_length()}}.
}
\examples{
diff_in_months(as.Date("2023-02-01"), as.Date("2022-02-01"))
}
