% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_in_time}
\alias{diff_in_time}
\title{Age difference in time}
\usage{
diff_in_time(x, y, units = "months")
}
\arguments{
\item{x}{Most recent date}

\item{y}{Least recent date}

\item{units}{Time units in which the time difference should be returned. Passed to \link[lubridate:time_length]{lubridate::time_length}.}
}
\value{
Absolute difference in months elapsed between \code{x} and \code{y}
}
\description{
Returns the absolute difference in time elapsed between two dates, as indicated by \code{\link[lubridate:time_length]{lubridate::time_length()}}.
}
\examples{
diff_in_time(as.Date("2023-02-01"), as.Date("2022-02-01"))
}
\author{
Gonzalo Garcia-Castro
}
