% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{download_surveys}
\alias{download_surveys}
\title{Download formR surveys}
\usage{
download_surveys(surveys, ...)
}
\arguments{
\item{surveys}{Name of the surveys in the formr run.}

\item{...}{Unused.}
}
\description{
Download formR surveys
}
\author{
Gonzalo Garcia-Castro
}
