% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{first_non_na}
\alias{first_non_na}
\title{First non-non-missing value}
\usage{
first_non_na(x)
}
\arguments{
\item{x}{Vector whose NAs will be replaced with first non-NA value}
}
\description{
First non-non-missing value
}
