% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_code_raw}
\alias{fix_code_raw}
\title{Fix raw codes}
\usage{
fix_code_raw(x)
}
\arguments{
\item{x}{Vector of \code{code} whose values should be fixed, based on \code{session}}
}
\description{
Fix raw codes
}
