% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_id_exp}
\alias{fix_id_exp}
\title{Fix id_exp}
\usage{
fix_id_exp(x)
}
\arguments{
\item{x}{Vector of \code{id_exp} whose values should be fixed}
}
\description{
Fix id_exp
}
\author{
Gonzalo Garcia-Castro
}
