% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_sex}
\alias{fix_sex}
\title{Fix sex (missing in first responses to BL-Lockdown)}
\usage{
fix_sex(x)
}
\arguments{
\item{x}{Vector of \code{sex} whose values should be fixed}
}
\description{
Fix sex (missing in first responses to BL-Lockdown)
}
