% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonology.R
\name{flatten_ipa}
\alias{flatten_ipa}
\title{Remove punctuation and fix non-ASCII characters from IPA transcriptions}
\usage{
flatten_ipa(x)
}
\arguments{
\item{x}{A character string with a phonological transcription in IPA format.}
}
\value{
A character string containing a phonological transcription in IPA format in which punctuation characters
have been removed.
}
\description{
Remove punctuation and fix non-ASCII characters from IPA transcriptions
}
\details{
Note that this function will effectively remove information about
syllabification and stress from the phonological representations.
}
\seealso{
\code{\link[=flatten_xsampa]{flatten_xsampa()}}, \code{\link[ipa:ipa]{ipa::ipa()}}
}
\author{
Gonzalo Garcia-Castro
}
