% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_age_bins}
\alias{get_age_bins}
\title{Select age bins flexibly}
\usage{
get_age_bins(x, width = 2)
}
\arguments{
\item{x}{Vector of ages (in months)}

\item{width}{Width of the age bins (in months, defaults to 2)}
}
\description{
Select age bins flexibly
}
