% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_time_stamp}
\alias{get_time_stamp}
\title{Get timestamps}
\usage{
get_time_stamp(x, cols, which = "first")
}
\arguments{
\item{x}{Data frame containing a column for the first time stamp and the
last time stamp of participants' responses the word inventory in each
language (Spanish and Catalan)}

\item{cols}{Character string vector indicating the name of the columns
containing the first and the last time stamps (in that order) of
participants' responses to a given language inventory.}

\item{which}{Which time stamp to consider:
\itemize{
\item \code{'first'} (by default)
\item \code{'last'}
}}
}
\description{
Get timestamps
}
\author{
Gonzalo Garcia-Castro
}
