% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Evaluate if x is included in y}
\usage{
x \%nin\% y
}
\arguments{
\item{x}{Vector whose elements will be evaluated}

\item{y}{Vector in which elements of \code{x} will be evaluated}
}
\description{
Evaluate if x is included in y
}
