% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_formr_short}
\alias{import_formr_short}
\title{Import short}
\usage{
import_formr_short(
  surveys = c("bilexicon_short_01_log", "bilexicon_short_02_welcome",
    "bilexicon_short_03_consent", "bilexicon_short_04_demo",
    "bilexicon_short_05_language", "bilexicon_short_06_words_catalan",
    "bilexicon_short_06_words_spanish"),
  ...
)
}
\arguments{
\item{surveys}{Name of formr surveys from the bilexicon_short run}

\item{...}{Unused.}
}
\description{
Import short
}
\author{
Gonzalo Garcia-Castro
}
