% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{launch_app}
\alias{launch_app}
\title{Launch bvq Shiny App in a browser}
\source{
https://github.com/gongcastro/bvq-app
}
\usage{
launch_app()
}
\description{
Launch bvq Shiny App in a browser
}
\details{
The BVQ Shiny App provides a visual interface to the bvq R package to explore the database. Its \href{https://github.com/gongcastro/bvq-app}{GitHub repository} contains the data, documentation, and R scripts needed to run the BVQ Shiny app.
}
