% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{merge_surveys}
\alias{merge_surveys}
\title{Merge surveys from the same formr run}
\usage{
merge_surveys(raw, participants_tmp, version)
}
\arguments{
\item{raw}{Raw survey data, as generated by \code{\link[=download_surveys]{download_surveys()}}}

\item{participants_tmp}{Participants dataset, inherited from inside the \verb{import_*()} function environment.}

\item{version}{character string indicating the name of the BVQ version being processed (must be \code{"BL-Lockdown"}, \code{"BL-Short"}, or \code{"BL-Long-2"}).}
}
\description{
Merge surveys from the same formr run
}
