% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pool.R
\docType{data}
\name{pool}
\alias{pool}
\title{Pool of words}
\format{
A data frame with 1601 rows and 20 variables:
\itemize{
\item item: item label, as indicated in the formr survey spreadsheets, items are unique within and across questionnaires.
\item language: language the item belongs to.
\item te: index associated to translation equivalents across languages.
\item label: item label, as presented to participants in the front-end of the questionnaire, some labels are not unique within or across questionnaires.
\item xsampa: phonological transcription in X-SAMPA format, transcribed from \code{ipa} using the \link[ipa:ipa]{ipa::ipa} function.
\item n_lemmas: an integer indicating the number of different lemmas showed in the item label to participants. for instance, the Spanish item \code{"spa_hierba"} was shown to in the questionnaire as \code{"hierba / césped"}. Lemma with similar roots were considered as one, such as the Spanish item \code{"spa_tonto"}, presented as \code{"tonto / tonta"} in the questionnaire.
\item is_multiword: an logical indicating  whether the item included a multi-word phrase as presented in the questionnaire. For instance the Spanish item \code{"spa_cepillodientes"} was shown as \code{"cepillo de dientes"} in the questionnaire, which includes three words.
\item subtlex_lemma: word label, as included in the corresponding version.
of SUBTLEX.
\item wordbank_lemma: word label, as indexed in Wordbank.
\item childes_lemma: word label, as it appears in the CHILDES English
corpora (based on \code{wordbank_lemma}).
\item semantic_category: semantic/functional category the items belongs to.
\item class: Functional category (verb, nouns, adjective, etc.).
\item version: what short version of the questionnaire does this item appear on?
\item include: should this item be included in analyses?
}
}
\usage{
pool
}
\description{
A dataset containing candidate words to be included in the questionnaires
with some lexical properties. Transcriptions were (a) generated manually, (b)
retrieved from \href{https://www.wiktionary.org/}{Wiktionary}. All transcriptions
have been manually double-checked and fixed if necessary.
}
\keyword{datasets}
