% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{process_survey}
\alias{process_survey}
\title{Process survey contents}
\usage{
process_survey(raw, participants_tmp, survey_name)
}
\arguments{
\item{raw}{Raw survey data, as generated by \code{\link[=download_surveys]{download_surveys()}}}

\item{participants_tmp}{Participants dataset, inherited from inside the \verb{import_*()} function environment.}

\item{survey_name}{character string indicating the name of the survey being processed (must be \code{"BL-Lockdown"}, \code{"BL-Short"}, or \code{"BL-Long-2"}).}
}
\description{
Process survey contents
}
\author{
Gonzalo Garcia-Castro
}
