% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop_adj}
\alias{prop_adj}
\title{Proportion, adjusted for zero- and one- inflation
\insertCite{gelman2020regression}{bvqdev}}
\usage{
prop_adj(x, n)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Number of tries}
}
\description{
Proportion, adjusted for zero- and one- inflation
\insertCite{gelman2020regression}{bvqdev}
}
\references{
\insertRef{gelman2020regression}{bvqdev}
}
