% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prop_adj_se}
\alias{prop_adj_se}
\title{Standard error of proportion, adjusted for zero- and one-inflation}
\usage{
prop_adj_se(x, n)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Number of trials}
}
\description{
Standard error of proportion, adjusted for zero- and one-inflation
}
\examples{
prop_adj_se(4, 60)
}
