% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_special_characters}
\alias{replace_special_characters}
\title{Replace special characters}
\usage{
replace_special_characters(x)
}
\arguments{
\item{x}{Character vector whose strings will be transformed}
}
\description{
Replace special characters
}
