% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonology.R
\name{syllabify_ipa}
\alias{syllabify_ipa}
\title{Syllabify phonological transcriptions in IPA or X-SAMPA formats.}
\usage{
syllabify_ipa(x, .sep = c("\\\\.", "'", "ˈ"))
}
\arguments{
\item{x}{A character string with a phonological transcription in IPA format.}

\item{.sep}{A vector of character strings indicating the characters that will be used to separate syllables. Takes \code{"\\\\."}, \code{"'"} and \code{"ˈ"} by default.}
}
\value{
A vector of characters in which each element is a syllable.
}
\description{
Syllabify phonological transcriptions in IPA or X-SAMPA formats.
}
\seealso{
\code{\link[=syllabify_xsampa]{syllabify_xsampa()}}, \code{\link[ipa:sampa]{ipa::xsampa()}}, \code{\link[ipa:ipa]{ipa::ipa()}}, \code{\link[=flatten_ipa]{flatten_ipa()}}
}
\author{
Gonzalo Garcia-Castro
}
