% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonology.R
\name{syllabify_xsampa}
\alias{syllabify_xsampa}
\title{Syllabify phonological transcriptions in X-SAMPA formats}
\usage{
syllabify_xsampa(x, .sep = c("\\\\.", "\\\\\\""))
}
\arguments{
\item{x}{A character string with a phonological transcription in X-SAMPA.}

\item{.sep}{A vector of character strings indicating the characters that will be used to separate syllables. Takes \code{"\\\\."} and \verb{"\\\\\\""} by default.}
}
\value{
A vector of characters in which each element is a syllable.
}
\description{
Syllabify phonological transcriptions in X-SAMPA formats
}
\examples{
\dontrun{
vct <- pool$xsampa[1:10]

syllabify_xsampa(vct)
}

}
\author{
Gonzalo Garcia-Castro
}
