% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonology.R
\name{syllable_str_ipa}
\alias{syllable_str_ipa}
\title{Get syllable structure from IPA phonological transcription}
\usage{
syllable_str_ipa(x, .sep = c("\\\\.", "\\\\\\""))
}
\arguments{
\item{x}{A character string with a phonological transcription in X-SAMPA format.}

\item{.sep}{Character separating syllables in the input transcriptions.}
}
\value{
A vector of characters in which each element is a syllable, in which vowels have been replaced with \code{"V"} and each consonants has been replaced with \code{"C"}.
}
\description{
Get syllable structure from IPA phonological transcription
}
\author{
Gonzalo Garcia-Castro
}
