% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocabulary.R
\name{vocab_dominance}
\alias{vocab_dominance}
\title{Compute dominance vocabulary (L1, L2)}
\usage{
vocab_dominance(x, ...)
}
\arguments{
\item{x}{Base data frame from which to calculate total vocabulary size}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}> Unquoted name of the
variable(s) to group data into. Vocabulary metrics will be calculated by
aggregating responses within the groups that result from the combination of
crossing of the variables provided in \code{...}. These variables can refer to
item properties (see \link{pool}, e.g., \code{semantic_category}) or to
participant properties (see \code{\link[=bvq_logs]{bvq_logs()}}, e.g., \code{lp}).}
}
\description{
Compute dominance vocabulary (L1, L2)
}
