% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vowels.R
\docType{data}
\name{vowels}
\alias{vowels}
\title{Inventory and classification of vowels in X-SAMPA and IPA formats.}
\format{
A data frame with 34 rows and 4 variables:
\itemize{
\item xsampa: phoneme symbol in \href{https://en.wikipedia.org/wiki/X-SAMPA}{X-SAMPA} format
\item ipa: phoneme symbol in \href{https://en.wikipedia.org/wiki/International_Phonetic_Alphabet}{IPA} format.
\item openness: \code{"Close"}, \code{"Near-close"}, \code{"Near-mid"}, \code{"Close-mid"}, \code{"Mid"}, \code{"Open-mid"}, \code{"Near-open"}, or \code{"Open"}. "An open vowel is a vowel sound in which the tongue is positioned as far as possible from the roof of the mouth. Open vowels are sometimes also called low vowels (in U.S. terminology) in reference to the low position of the tongue.
\item frontness: \verb{"Back}, \verb{"Central}, or \verb{"Front}. "A front vowel is a class of vowel sounds used in some spoken languages, its defining characteristic being that the highest point of the tongue is positioned as far forward as possible in the mouth without creating a constriction that would otherwise make it a consonant. Front vowels are sometimes also called bright vowels because they are perceived as sounding brighter than the back vowels.
}
}
\source{
https://en.wikipedia.org/wiki/X-SAMPA
}
\usage{
vowels
}
\description{
A dataset containing most vowel phonemes identified by the \href{https://en.wikipedia.org/wiki/International_Phonetic_Association}{International Phonetic Association} (IPA). Phonemes are classified across three dimensions: place of articulation, manner of articulation, and voicing. Each phoneme is assigned a symbol in X-SAMPA and IPA format.
}
\keyword{datasets}
