% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{bvq_connect}
\alias{bvq_connect}
\title{Authenticate in Google and formr}
\usage{
bvq_connect(google_email = NULL, verbose = TRUE, password = NULL)
}
\arguments{
\item{google_email}{E-mail used in Google Drive account. If NULL (default), it is assumed to be the same as formr_email.}

\item{verbose}{Should progress messages and warnings be printed in the console}
}
\value{
Logical. TRUE if Google and formr authentication was successful, FALSE if authentication of any of the two failed.
}
\description{
Authenticate in Google and formr
}
\details{
This function tries to log in to the formr API with the user-provided password (argument \code{password}) or retrieving it from the global environment (FORMR_PWD in .Renviron)
}
