% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{bvq_connect}
\alias{bvq_connect}
\title{Authenticate in Google and formr}
\usage{
bvq_connect(google_email = NULL, password = NULL)
}
\arguments{
\item{google_email}{E-mail used in Google Drive account. If \code{NULL} (default),
it is assumed to be the same as \code{formr_email}.}

\item{password}{Character string with the password to formr (\code{NULL} by
default).}
}
\value{
Logical. \code{TRUE} if Google and formr authentication was successful,
\code{FALSE} if authentication of any of the two failed.
}
\description{
This function tries to log in to the formr API with the user-provided
password (argument \code{password}) or retrieving it from the global environment
(\code{FORMR_PWD} in .Renviron)
}
