% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{bvq_norms}
\alias{bvq_norms}
\title{Generate item-level norms for age, sex, language profile and item dominance}
\usage{
bvq_norms(
  participants = bvq_participants(),
  responses = bvq_responses(participants),
  ...,
  te = NULL,
  item = NULL,
  age = c(0, Inf)
)
}
\arguments{
\item{participants}{Participants data frame, as generated by
\code{\link[=bvq_participants]{bvq_participants()}}}

\item{responses}{Responses data frame, as generated by \code{\link[=bvq_responses]{bvq_responses()}}.}

\item{...}{<\code{\link[rlang:dyn-dots]{dynamic-dots}}>. Unquoted name of the variable(s) to group
data into. Norms will be calculated by aggregating responses
within the groups that result from the combination of crossing of the
variables provided in \code{...}. These variables can refer to item properties
(see \link{pool}, e.g., \code{semantic_category}) or to participant properties (see
\code{\link[=bvq_logs]{bvq_logs()}}, e.g., \code{lp}).}

\item{te}{Translation equivalent for which the norms should be computed.
\itemize{
\item If NULL (default), norms are computed exclusively for the items indicated in \code{item}.
\item If TRUE, norms are computed for both the item indicated in the \code{item} argument, and for its translation.
\item If FALSE, norms are computed exclusively for the items indicated in \code{item} (same as NULL)
\item If numeric vector, norms are computed for all items corresponding to the translation equivalents indicated in this argument.
}}

\item{item}{Character string indicating the item to compute norms for. If
left \code{NULL} (default) norms will be computed for all items. You can check
the available items in the \link{pool} data set running \code{data("pool")}.}

\item{age}{Numeric vector of length two (\emph{min}-\emph{max}) indicating the age range of
participants to compute norms for.}
}
\value{
A data frame (actually, a \link[tibble:tibble]{tibble::tibble} with the proportion of
participants in the sample that understand or produce the items indicated
in \code{item}, or the translation equivalents indicated in \code{te}.
The output contains the following variables:
\itemize{
\item te: an integer identifying the translation equivalent (a.k.a., pair of cross-language synonyms, doublets) the item belongs to.
\item item: character string indicating the item identifier (e.g., \code{spa_mesa}). This value is unique for each item. Responses to the same item from different participants are linked by the same \code{item} value.
\item language: a character string indicating the language the item response belongs to: \code{"Catalan"} if item in Catalan), \code{"Spanish"} if item in Spanish.
\item age: an numeric vector of length 1 or 2 indicating the age range of participants (in months) for which the estimates should be computed. If a non-integer is provided (e.g., \code{15.36}, it is rounded downwards using \code{\link[=floor]{floor()}}.)
\item type: a character string indicating the vocabulary type computed: \code{"understands"} if option 'Understands' was selected, and \code{"produces"} if option 'Understands & Says' was selected.
\item item_dominance: a character string that takes the value \code{"L1"} if the item belongs to participants' language of most exposure, and L2 if the item belongs to participants' language of least exposure.
\item label: a character string indicating the text presented to participants in the questionnaire (replacing the \code{item} identifier).
\item .sum: a positive integer indicating the number of positive responses: \code{responses} is 2 (Understands) or 3 (Understands & Says) for \code{type = "understands"}, and 3 (Understands & Says) if \code{type = "produces"}.
\item .n: a positive integer indicating the total number number of responses (useful for computing proportions).
\item .prop: a numeric value ranging from 0 to 1 (both included) indicating the estimated proportion of participants that provided a positive response, adjusted following Gelman et al.'s method to account for zero- and one-inflation (see function \link{prop_adj}).
Additionally, any variables specified in the \code{.by} argument are preserved as grouping variables.
}
}
\description{
This function generates a data frame with the estimated proportion of
children that understand and/or produce some items for a selected age range
and participant profiles. Estimated proportions and corresponding standard
errors and confidence intervals are computed adjusting for zero- and
one-inflation (see function \code{\link[=prop_adj]{prop_adj()}}).
}
\examples{
\dontrun{
responses <- bvq_responses()

bvq_norms(
  participants = participants,
  responses = responses,
  item = "cat_casa",
  age = c(22, 22),
  lp
)

my_items <- c("cat_gos", "cat_gat")

bvq_norms(
  participants = participants,
  responses = responses,
  item = my_items,
  te = TRUE,
  age = c(15, 16)
)
}

}
\author{
Gonzalo Garcia-Castro
}
