% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{bvq_responses}
\alias{bvq_responses}
\title{Retrieve and update local and/or remote data from formr}
\usage{
bvq_responses(participants = bvq_participants())
}
\arguments{
\item{participants}{Participants data frame, as generated by
\code{\link[=bvq_participants]{bvq_participants()}}. If \code{NULL} (default), \code{\link[=bvq_participants]{bvq_participants()}} is run.}
}
\value{
A data frame (actually, a \link[tibble:tibble]{tibble::tibble} containing participant's
responses to each item, along with some session-specific metadata. The
output includes the following variables:
\itemize{
\item child_id: a character string with five digits indicating a participant's identifier in the database from the \href{https://www.upf.edu/web/cbclab}{Laboratori de Recerca en Infància} at Universitat Pompeu Fabra. This value is always the same for each participant, so that different responses from the same participant share the same \code{child_id}.
\item response_id: a character string identifying a single response to the questionnaire. This value is always unique for each response to the questionnaire, even for responses from the same participant.
\item time: a numeric value indicating how many times a given participant has been sent the questionnaire, regardless of whether they completed it or not.
\item version: a character string indicating what version of the questionnaire a given participant filled in. Different versions may contain a different subset of items, and the administration instructions might vary slightly (see formr questionnaire templates in the \href{https://github.com/gongcastro/bvq}{GitHub repository}). Also, different versions were designed, implemented, and administrated at different time points (e.g., before/during/after the COVID-related lockdown).
\item version_list: a character string indicating the specific list of
items a participant was assigned to. Only applies in the case of short
versions of BVQ, such as bvq-short, bvq-long, bvq-lockdown, or bvq-1.0.0, where the
list of items was partitioned into several versions.#' * item: character string indicating the item identifier (e.g., \code{spa_mesa}). This value is unique for each item. Responses to the same item from different participants are linked by the same \code{item} value.
\item response: integer indicating the participant's response to a give item: \code{1} if \code{"No"} (the participant does not understand or produce the word), \code{2} if "Understands" (the participants understands the word), or \code{3} if "Understands and Says" (the participant understands and produces the item).
\item date_birth: \link[lubridate:date_utils]{lubridate::Date} indicating participants birth date.
\item date_started: \link[lubridate:date_utils]{lubridate::Date} indicating when participants logged to the questionnaire for the first time.
\item date_finished: \link[lubridate:date_utils]{lubridate::Date} indicating when participants logged to the questionnaire for the last time.
\item sex: a character string indicating participants' biological sex, as reported by the parents.
\item doe_spanish: a numeric value ranging from 0 to 1 indicating participants' daily exposure to Spanish, as estimated by parents/caregivers This value aggregates participants' exposure to any variant of Spanish (e.g., European and American Spanish).
\item doe_catalan: a numeric value ranging from 0 to 1 indicating participants' daily exposure to Catalan, as estimated by parents/caregivers This value aggregates participants' exposure to any variant of Catalan (e.g., Catalan from Majorca or Barcelona).
\item doe_others: a numeric value ranging from 0 to 1 indicating participants' daily exposure to languages other than Spanish or Catalan, as estimated by parents/caretakers, aggregating participants' exposure to all those other languages (e.g., Norwegian, Arab, Swahili).
\item edu_parent1: a character string indicating the educational attainment of one of the parents/caretakers.
\item edu_parent2: a character string indicating the educational attainment of the other parent/caretaker, if any.
}
}
\description{
This function generates a data frame with participant's responses to each
item, along with some session-specific metadata. It takes \code{participants}
(the output of \code{\link[=bvq_participants]{bvq_participants()}}) and \code{runs} (a character vector that can
take zero, one, or multiple of the following values: \code{"formr2"},
\code{"formr-short"}, \code{"formr-lockdown"}) as arguments.
}
\examples{
\dontrun{
bvq_responses()
}

}
\author{
Gonzalo Garcia-Castro
}
