% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/responses.R
\name{bvq_responses}
\alias{bvq_responses}
\title{Retrieve and update local and/or remote data from formr}
\usage{
bvq_responses(
  participants = NULL,
  runs = c("BL-Long2", "BL-Lockdown"),
  longitudinal = "all",
  ...
)
}
\arguments{
\item{participants}{Participants data frame, as generated by
\code{\link[=bvq_participants]{bvq_participants()}}. If \code{NULL} (default), \code{\link[=bvq_participants]{bvq_participants()}} is run.}

\item{runs}{formr runs to update}

\item{longitudinal}{Should longitudinal data be included? If \code{"all"}
(default), all responses (including repeated measures) are included. If
"no", participants with more than one responses to the questionnaire
(regardless of the version) are excluded. If \code{"first"}, only the first
response of each participant is included. If \code{"last"}, only the last
response of each participant is included. If \code{"only"}, only responses with
repeated measures are included.}

\item{...}{Extra arguments passed to \link{bvq_connect}}
}
\value{
A data frame (actually, a \link[tibble:tibble]{tibble::tibble} containing participant's
responses to each item, along with some session-specific metadata. The
output includes the following variables:
\itemize{
\item id: a character string indicating a participant's identifier. This value is always the same for each participant, so that different responses from the same participant share the same \code{id}.
\item id_exp: a character string indicating a participant's identifier in the context of the particular study in which the participant was tested and invited to fill in the questionnaire. This value is always the same for each participant within the same study, so that different responses from the same participant in the same study share \code{id_exp}. The same participant may have different \code{id_exp} across different studies.
\item id_db: a character string with five digits indicating a participant's identifier in the database from the \href{https://www.upf.edu/web/cbclab}{Laboratori de Recerca en Infància} at Universitat Pompeu Fabra. This value is always the same for each participant, so that different responses from the same participant share the same \code{id_db}.
\item time: a numeric value indicating how many times a given participant has been sent the questionnaire, regardless of whether they completed it or not.
\item code: a character string identifying a single response to the questionnaire. This value is always unique for each response to the questionnaire, even for responses from the same participant.
\item study: a character string indicating the study in which the participant was invited to fill in the questionnaire. Frequently, participants that filled in the questionnaire came to the lab to participant in a study, and were then invited to fill in the questionnaire later. This value indicates what study each participant was tested in before being sent the questionnaire.
\item version: a character string indicating what version of the questionnaire a given participant filled in. Different versions may contain a different subset of items, and the administration instructions might vary slightly (see formr questionnaire templates in the \href{https://github.com/gongcastro/bvqdev}{GitHub repository}). Also, different versions were designed, implemented, and administrated at different time points (e.g., before/during/after the COVID-related lockdown).
\item time_stamp: a date value (see lubridate package) in \code{yyyy/mm/dd} format indicating the date in which participants responded to the last item of their questionnaire response. Note: some participants took longer to complete the questionnaire since they started filling items in.
\item language: a character string indicating the language the item response belongs to: \code{"Catalan"} if item in Catalan), \code{"Spanish"} if item in Spanish.
\item item: character string indicating the item identifier (e.g., \code{spa_mesa}). This value is unique for each item. Responses to the same item from different participants are linked by the same \code{item} value.
\item response: integer indicating the participant's response to a give item: \code{1} if \code{"No"} (the participant does not understand or produce the word), \code{2} if "Understands" (the participants understands the word), or \code{3} if "Understands and Says" (the participant understands and produces the item).
\item date_birth: a date value (see lubridate package) in \code{yyyy/mm/dd} format indicating participants birth date.
\item age: a numeric value indicating the number of months elapsed since participants' birth date until they filled in the last item of their questionnaire response.
\item sex: a character string indicating participants' biological sex, as reported by the parents.
\item postcode: a character string indicating participants' household postcode.
\item edu_parent1: a character string indicating the educational attainment of one of the parents/caretakers.
\item edu_parent2: a character string indicating the educational attainment of the other parent/caretaker, if any.
\item doe_spanish: a numeric value ranging from 0 to 1 indicating participants' daily exposure to Spanish, as estimated by parents/caretakers. This value aggregates participants' exposure to any variant of Spanish (e.g., European and American Spanish).
\item doe_catalan: a numeric value ranging from 0 to 1 indicating participants' daily exposure to Catalan, as estimated by parents/caretakers. This value aggregates participants' exposure to any variant of Catalan (e.g., Catalan from Majorca or Barcelona).
\item doe_others: a numeric value ranging from 0 to 1 indicating participants' daily exposure to languages other than Spanish or Catalan, as estimated by parents/caretakers, aggregating participants' exposure to all those other languages (e.g., Norwegian, Arab, Swahili).
\item randomisation: a character string indicating the specific list of items a participant was assigned to. Only applies in the case of short versions of BVQ, such as BL-Short, BL-Short-2 or BL-Lockdown, where the list of items was partitioned into several versions.
\item doe_spanish_lockdown: a numeric value ranging from 0 to 1 indicating participants' daily exposure to Spanish during the COVID-19 lockdown, as estimated by parents/caretakers. This value aggregates participants' exposure to any variant of Spanish (e.g., European and American Spanish).
\item doe_catalan_lockdown: a numeric value ranging from 0 to 1 indicating participants' daily exposure to Catalan during the COVID-19 lockdown, as estimated by parents/caretakers. This value aggregates participants' exposure to any variant of Catalan (e.g., Catalan from Majorca or Barcelona).
\item doe_others_lockdown: a numeric value ranging from 0 to 1 indicating participants' daily exposure to languages other than Spanish or Catalan during the COVID-19 lockdown, as estimated by parents/caretakers, aggregating participants' exposure to all those other languages (e.g., Norwegian, Arab, Swahili).
\item dominance: a character string indicating the language of highest exposure (\code{"Catalan"} or \code{"Spanish"}), as reported by parents. If exposure is identical for both language, \code{"Catalan"} is assigned.
}
}
\description{
This function generates a data frame with participant's responses to each
item, along with some session-specific metadata. It takes \code{participants}
(the output of \code{\link[=bvq_participants]{bvq_participants()}}) and \code{runs} (a character vector that can
take zero, one, or multiple of the following values: \code{"formr2"},
\code{"formr-short"}, \code{"formr-lockdown"}) as arguments. Only responses from the
versions indicated in \code{runs} input will be updated.
}
\author{
Gonzalo Garcia-Castro
}
