% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvqdev-package.R
\docType{package}
\name{bvqdev-package}
\alias{bvqdev}
\alias{bvqdev-package}
\title{bvqdev: Barcelona Vocabulary Questionnaire Database and Helper Functions}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Download, clean, and process the Barcelona Vocabulary Questionnaire data. BVQ is a vocabulary inventory developed for assesing the vocabulary of Catalan-Spanish bilinguals infants from the Metropolitan Area of Barcelona (Spain), aged 10 to 40 months. This package includes functions to download the data from formr servers, and return the processed data in multiple formats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gongcastro.github.io/bvqdev/}
  \item \url{https://github.com/gongcastro/bvqdev/}
  \item Report bugs at \url{https://github.com/gongcastro/bvqdev/issues}
}

}
\author{
\strong{Maintainer}: Gonzalo Garcia-Castro \email{gonzalo.garciadecastro@upf.edu} (\href{https://orcid.org/0000-0002-8553-4209}{ORCID})

}
\keyword{internal}
