% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bvqdev-package.R
\docType{package}
\name{bvqdev-package}
\alias{bvqdev}
\alias{bvqdev-package}
\title{bvqdev: What the Package Does (One Line, Title Case)}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package is intended for internal use by the mantainers of the bvq package. It contains functions that download the BVQ survey data from the formR servers, where the questionnaires are hosted and the data are stored, clean them, and upload them to the Universitat Pompeu Fabra servers. The final data is what users of the bvq package can retrieve.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gongcastro.github.io/bvqdev/}
  \item Report bugs at \url{https://github.com/gongcastro/multilex/issues}
}

}
\author{
\strong{Maintainer}: Gonzalo Garcia-Castro \email{gonzalo.garciadecastro@upf.edu} (\href{https://orcid.org/0000-0002-8553-4209}{ORCID})

}
\keyword{internal}
