% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms.R
\name{check_arg_te}
\alias{check_arg_te}
\title{Check argument \code{te} in \code{\link[=bvq_norms]{bvq_norms()}} function}
\usage{
check_arg_te(te, item)
}
\description{
Check argument \code{te} in \code{\link[=bvq_norms]{bvq_norms()}} function
}
