% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonology.R
\name{check_xsampa}
\alias{check_xsampa}
\title{Check that all characters included in X-SAMPA phonological transcriptions are part of the X-SAMPA alphabet}
\usage{
check_xsampa(x)
}
\arguments{
\item{x}{A vector of character strings with at least one element that contains
phonological transcriptions in X-SAMPA format.}
}
\value{
A logical (invisible) value indicating whether all symbols in \code{x} are part of the X-SAMPA alphabet.
}
\description{
Check that all characters included in X-SAMPA phonological transcriptions are part of the X-SAMPA alphabet
}
