% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{diff_in_months}
\alias{diff_in_months}
\title{Age difference in months}
\usage{
diff_in_months(x, y)
}
\arguments{
\item{x}{Most recent date}

\item{y}{Least recent date}
}
\description{
Age difference in months
}
