% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phonology.R
\name{flatten_xsampa}
\alias{flatten_xsampa}
\title{Remove punctuation from SAMPA transcriptions}
\usage{
flatten_xsampa(x)
}
\arguments{
\item{x}{A character string with a phonological transcription in X-SAMPA format.}
}
\value{
A character string containing a phonological transcription in X-SAMPA format in which punctuation characters
have been removed.
}
\description{
Remove punctuation from SAMPA transcriptions
}
\details{
Note that this function will effectively remove information about
syllabification and stress from the phonological representations.
}
\author{
Gonzalo Garcia-Castro
}
