% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_doe}
\alias{get_doe}
\title{Summarise language profile}
\usage{
get_doe(x, languages)
}
\arguments{
\item{x}{Data frame that contains each degree of exposure as columns,
named \verb{language_doe_*}.}

\item{languages}{Character vector of languages to compute degree of exposure
for (all others will be considered as \code{doe_others}).}
}
\description{
Summarise language profile
}
