% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_doe}
\alias{get_doe}
\title{Summarise language profile}
\usage{
get_doe(...)
}
\arguments{
\item{...}{Character vector of languages to compute degree of exposure
for (all others will be considered as \code{doe_others}).}
}
\description{
Summarise language profile
}
