% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_longitudinal}
\alias{get_longitudinal}
\title{Deal with repeated measures}
\usage{
get_longitudinal(x, longitudinal = "all")
}
\arguments{
\item{x}{A data frame containing a column for participants (each participant
gets a unique ID), and a column for times (a numeric value indicating how
many times each participant appears in the data frame counting this one).
One participant may appear several times in the data frame, with each time
with a unique value of \code{time}.}

\item{longitudinal}{A character string indicating what subset of the
participants should be returned:
\itemize{
\item \code{"all"} (default) returns all participants.
\item \code{"no"} remove all participants with more than one response.
\item \code{"only"} returns only participants with more than one response in the
dataset (i.e., longitudinal participants).
\item \code{"first"} returns the first response of each participant (participants with only one appearance are
included).
\item \code{"last"} returns the last response from each participant (participants with only one response are included).
}}
}
\value{
A subset of the data frame \code{x} with only the selected cases,
according to \code{longitudinal}.
}
\description{
Deal with repeated measures
}
\examples{
id <- c(1, 1, 1, 2, 2, 3, 4, 4, 4, 4, 5, 6, 7, 7, 8, 9, 10, 10)
sums <- rle(sort(id))[["lengths"]]
dat <- data.frame(id, time = unlist(sapply(sums, function(x) seq(1, x))))
get_longitudinal(dat, "first")
}
\author{
Gonzalo Garcia-Castro
}
