% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_time_stamp}
\alias{get_time_stamp}
\title{Get timestamps}
\usage{
get_time_stamp(..., which = "first")
}
\arguments{
\item{...}{Columns containing the first and the last time stamps (in that order) of
participants' responses to a given language inventory.}

\item{which}{Which time stamp to consider:
\itemize{
\item \code{'first'} (by default)
\item \code{'last'}
}}
}
\value{
A datetime vector with the first or last of each time stamp
}
\description{
Get timestamps
}
\examples{
library(dplyr)
x <- data.frame(start = as.Date(c("2023-02-01", "2023-02-20")),
                finish = as.Date(c("2023-03-15", "2023-02-21")))

y <- mutate(x, 
            time_stamp_default = get_time_stamp(start, finish), 
            time_stamp_first = get_time_stamp(start, finish, which = "first"), 
            time_stamp_last = get_time_stamp(start, finish, which = "last"))
(y)

}
\author{
Gonzalo Garcia-Castro
}
