% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_formr_lockdown}
\alias{import_formr_lockdown}
\title{Import lockdown data}
\usage{
import_formr_lockdown(
  surveys = c("bilexicon_lockdown_01_log", "bilexicon_lockdown_02_welcome",
    "bilexicon_lockdown_03_consent", "bilexicon_lockdown_04_demo",
    "bilexicon_lockdown_05_language", "bilexicon_lockdown_06_words_catalan",
    "bilexicon_lockdown_06_words_spanish"),
  ...
)
}
\arguments{
\item{surveys}{Name of formr surveys from the \code{bilexicon_lockdown} run.}

\item{...}{Unused.}
}
\description{
Import lockdown data
}
\author{
Gonzalo Garcia-Castro
}
