% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{prop_adj}
\alias{prop_adj}
\title{Proportion, adjusted for zero- and one-inflation}
\usage{
prop_adj(x, n)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Number of tries}
}
\value{
A numeric scalar.
}
\description{
Proportion, adjusted for zero- and one-inflation
}
\details{
It is very common that a large proportion of the participants know or do not know some word.
Vocabulary sizes and word prevalence norms in package are calculated using an estimate that
adjusts for zero- and one-inflation so that, at the population level such estimates are more
likely to be accurate.
}
\examples{
prop_adj(4, 60)

}
