% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{prop_adj}
\alias{prop_adj}
\title{Proportion, adjusted for zero- and one- inflation}
\usage{
prop_adj(x, n)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Number of tries}
}
\description{
Proportion, adjusted for zero- and one- inflation
}
\examples{
prop_adj_se(4, 60)
}
