% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{prop_adj_ci}
\alias{prop_adj_ci}
\title{Confidence interval of proportion, adjusted for zero- and one-inflation}
\usage{
prop_adj_ci(x, n, .width = 0.95)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Number of tries}

\item{.width}{Confidence level (defaults to .95)}
}
\description{
Confidence interval of proportion, adjusted for zero- and one-inflation
}
\examples{
prop_adj_ci(4, 60, .width = 0.89)
}
